refactor: relocate inverse transform
