+++ LGHackerton/preprocess/__init__.py
@@ -6,6 +6,10 @@ from LGHackerton.preprocess.preprocess_pipeline_v1_1 import (
     L,
     H,
 )
+from LGHackerton.preprocess.transforms import (
+    symmetric_transform,
+    inverse_symmetric_transform,
+)
 
 __all__ = [
     "Preprocessor",
@@ -14,4 +18,6 @@ __all__ = [
     "SALES_COL",
     "L",
     "H",
+    "symmetric_transform",
+    "inverse_symmetric_transform",
 ]
