--- src/__init__.py
@@ -1,1 +1,0 @@
-"""Top-level package for project-specific utilities."""
