--- src/data/__init__.py
@@ -1,1 +1,0 @@
-"""Data utilities and preprocessing helpers."""
